# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/test/order/update": {
            "get": {
                "description": "更新订单状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test_order"
                ],
                "summary": "更新订单状态",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_handler.updateOrderStateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/delete": {
            "delete": {
                "description": "删除一个账户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "删除账户",
                "parameters": [
                    {
                        "description": "删除信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "登录一个账户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "登出一个账户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "登出",
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "注册一个新的账户",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "注册账户",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.registerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户id",
                        "schema": {
                            "$ref": "#/definitions/handler.userIdReq"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/user/update": {
            "put": {
                "description": "修改一个账户的密码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.errorReturn": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handler.registerReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "handler.updateReq": {
            "type": "object",
            "required": [
                "new_password",
                "old_password",
                "username"
            ],
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.userIdReq": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "test_handler.updateOrderStateReq": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}